refactor: 사용자 정의 예외를 만들어 적용하고 핸들러를 리팩토링 합니다.
