+++ src/main/java/nextstep/subway/exception/ErrorResponse.java
@@ -1,14 +1,13 @@
-package nextstep.subway.common;
+package nextstep.subway.exception;
 
+import lombok.Getter;
+
+@Getter
 public class ErrorResponse {
 
     private final String message;
 
     public ErrorResponse(final String message) {
         this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
     }
 }
